+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 
 
+## [2.1.0](https://github.com/supercharge/arrays/compare/v2.0.0...v2.1.0) - 2022-02-xx
+
+### Added
+- `reduce()` method: runs a given reducer function on each item in the array and passes the accumulator to the next iteration
+- `groupBy()` method: group the array by a given key
+
+
 ## [2.0.0](https://github.com/supercharge/arrays/compare/v1.2.0...v2.0.0) - 2022-01-21
 
 ### Added